Proyek Akhir (Javadoc + fixes and adjustments)

+++ src/main/java/com/armondHarerJSleepJS/Rating.java
@@ -1,13 +1,16 @@
 package com.armondHarerJSleepJS;
 
-
+/**
+ * Class untuk rating dalam JSleep
+ * @author Armond Harer
+ */
 public class Rating
 {
     private long count;
     private long total;
 
     /**
-     * Constructor for objects of class rating
+     * Konfigurasi awal rating ruangan 
      */
     public Rating()
     {
@@ -15,12 +18,20 @@ public class Rating
         this.count = 0;
     }
 
+    /**
+     * @param rating | rating dari pelanggan ditambah ke total
+     */
     public void insert(int rating)
     {
         this.total += rating;
         this.count++;
     }
     
+    /**
+     * 
+     * @param count | jumlah rating
+     * @return rating rata-rata dari ruangan 
+     */
     public double getaverage(int count)
     {
         if (count <= 0)
@@ -30,20 +41,29 @@ public class Rating
         }
         else
         {
-            return this.total/this.count;
+            return (double)this.total/(double)this.count;
         }
     }
     
+    /**
+     * @return getter untuk jumlah rating
+     */
     public long getcount()
     {
         return this.count;
     }
     
+    /**
+     * @return getter untuk rating total
+     */
     public long gettotal()
     {
         return this.total;
     }
     
+    /**
+     * @return informasi rating dalam bentuk string
+     */
     public String toString(){
         return "Rating Tally\nTotal: " + total + "\nCount: " + count;
     }
